Fix for client NPE issue - don't missend player spawn packets (#1525)
* Fix #1360 - don't missend player spawn packets

Player spawn packets were improperly sent by the Minecraft internals in early NPC spawn sequence, when they are not valid to be sent yet. This patch blocks their sending until the tracker system is pushed onto the NPC. Tested and functional.

* improve long name cutting order, fixes #1422

ensures that root name will always be 16 characters if the given name was too long - previously, the root name could be as small as a single character (for input name of 17 characters).